Remove Cmake static library workaround
In order to build the static version of the library, simply pass Cmake
-DBUILD_SHARED_LIBS=OFF.

A single Cmake build is intended to only install EITHER shared or static
libraries depending on the BUILD_SHARED_LIBS option.

This changes the default from forcing shared libs to using the default
of building static libs.